+++ bus_user/serial_derivatives.py
@@ -1,17 +1,17 @@
 from typing import *
 
 
-from .serial_client import SerialClient, AddressAutoAcceptVariant
+from .serial_client import SerialClient, Type__AddressAutoAcceptVariant
 from .serial_server import SerialServer_Base, SerialServer_Example
 
 
 # =====================================================================================================================
 class SerialClient_Shorted(SerialClient):
-    ADDRESS = AddressAutoAcceptVariant.FIRST_FREE__SHORTED
+    ADDRESS = Type__AddressAutoAcceptVariant.FIRST_FREE__SHORTED
 
 
 class SerialClient_Emulated(SerialClient):
-    ADDRESS = AddressAutoAcceptVariant.FIRST_FREE__PAIRED_FOR_EMU
+    ADDRESS = Type__AddressAutoAcceptVariant.FIRST_FREE__PAIRED_FOR_EMU
     _EMULATOR__CLS = SerialServer_Example
     _EMULATOR__START = True